Share delegated token test helper with GCP PR
The inline delegatedTokenTestContext copy in azure_test.go collides
with the identically-named helper added by the GCP delegated auth PR
(#4590) - whichever PR merges second would break the tests module
compile on master. Restore the shared provider-generic helper file
from the GCP branch instead: identical content on both branches means
the second merge is a no-op for this path.
